HW05. Ability to register new books.

+++ hw05/src/main/java/ru/otus/spring/sokolovsky/hw05/domain/AuthorDao.java
@@ -3,7 +3,7 @@ package ru.otus.spring.sokolovsky.hw05.domain;
 import java.util.List;
 
 public interface AuthorDao {
-    Author getById(int id);
+    Author getById(long id);
 
     List<Author> getAll();
