+++ hw05/src/main/java/ru/otus/spring/sokolovsky/hw05/domain/GenreDao.java
@@ -3,7 +3,7 @@ package ru.otus.spring.sokolovsky.hw05.domain;
 import java.util.List;
 
 public interface GenreDao {
-    Genre getById(int id);
+    Genre getById(long id);
 
     List<Genre> getAll();
