+++ hw05/src/main/java/ru/otus/spring/sokolovsky/hw05/jdbcdao/AuthorDaoImpl.java
@@ -23,7 +23,7 @@ public class AuthorDaoImpl extends BaseDao implements AuthorDao {
     }
 
     @Override
-    public Author getById(int id) {
+    public Author getById(long id) {
         return jdbcTemplate.queryForObject(
                 createSelectBuilder().useFilterFields("id").toString(),
                 new HashMap<>() {{
